Customize Gradle Asciidoctor build task to properly identify doc source.
Other Spring Build Conventions based projects assume includes are in _includes but SBDG has includes outside of _includes. This means we must change the sources to only be 'index.adoc'.

Update docs build to configure every task of type AbstractAsciidoctorTask so that asciidoctorPdf picks up the custom attributes set and the updated sources.