Remove Absolute Path to stylesheet
An absolute path breaks the css for a few reasons:

a) Locally it is missing because everything is copied into a
temporary directory for processing. The temp directory exists when the
html is written out, but is removed once processing is done. The output
is still referring to the styles in the temp dir after processing is done.
b) When the html is deployed it will refer to the styles on the local
directory that the docs were built. When other users view it on the web
server, the dir will not exist.

This commit removes the stylesdir definition to fix this.