Ensure Asciidoctor attributes are configured for PDFs too
In previous versions of asciidoctor gradle plugin asciidoctor was an
extension that was read by all asciidoctor tasks. This meant that
configuring the following impacted all asciidoctor tasks (including pdf)

asciidoctor {
   attributes ...
}

The latest version of asciidoctor moved conventions to asciidoctorj. The
reason for this is there was a task named asciidoctor and a convention
named asciidoctor which caused confusion for users on which was being
configured.

This commit switches to the new asciidoctorj convention/extension so that
attributes are configured for all asciidoctor tasks (i.e. both the
asciidoctor and asciidoctorPdf tasks)